change in tag backoffice controller for add picture upload

+++ src/Controller/Backoffice/TagController.php
@@ -5,6 +5,7 @@ namespace App\Controller\Backoffice;
 use App\Entity\Tag;
 use App\Form\TagType;
 use App\Repository\TagRepository;
+use App\Service\PictureService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -23,13 +24,19 @@ class TagController extends AbstractController
     }
 
     #[Route('/new', name: 'app_tag_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, PictureService $pictureService): Response
     {
         $tag = new Tag();
         $form = $this->createForm(TagType::class, $tag);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $pictureFile = $form->get('pictures')->getData();
+            if ($pictureFile) {
+                $relativePath = $pictureService->add($pictureFile, 'tag');
+                $tag->setPicture($relativePath);
+            }
+
             $entityManager->persist($tag);
             $entityManager->flush();
 
@@ -38,7 +45,7 @@ class TagController extends AbstractController
 
         return $this->render('backoffice/tag/new.html.twig', [
             'tag' => $tag,
-            'form' => $form,
+            'form' => $form->createView(),
         ]);
     }
 
@@ -51,12 +58,27 @@ class TagController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_tag_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Tag $tag, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Tag $tag, EntityManagerInterface $entityManager, PictureService $pictureService): Response
     {
         $form = $this->createForm(TagType::class, $tag);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            // Gérer le téléchargement de l'image
+            $pictureFile = $form->get('pictures')->getData();
+
+            if ($pictureFile) {
+                // Supprimer l'ancienne image si elle existe
+                if ($tag->getPicture()) {
+                    $pictureService->delete($tag->getPicture());
+                }
+
+                // Ajouter la nouvelle image
+                $relativePath = $pictureService->add($pictureFile, 'tag');
+                $tag->setPicture($relativePath);
+            }
+
+            $entityManager->persist($tag);
             $entityManager->flush();
 
             return $this->redirectToRoute('app_tag_index', [], Response::HTTP_SEE_OTHER);
@@ -64,18 +86,30 @@ class TagController extends AbstractController
 
         return $this->render('backoffice/tag/edit.html.twig', [
             'tag' => $tag,
-            'form' => $form,
+            'form' => $form->createView(),
         ]);
     }
 
     #[Route('/{id}', name: 'app_tag_delete', methods: ['POST'])]
-    public function delete(Request $request, Tag $tag, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, tag $tag, EntityManagerInterface $entityManager, PictureService $pictureService): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $tag->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($tag);
-            $entityManager->flush();
-        }
+        if ($this->isCsrfTokenValid('delete' . $tag->getId(), $request->request->get('_token'))) {
+            // Supprimer l'image associée si elle existe
+            if ($tag->getPicture()) {
+                $relativePath = $tag->getPicture();
+                if ($pictureService->delete($relativePath)) {
+                } else {
+                    $this->addFlash('danger', 'La suppression de l\'image a échoué');
+                }
+            }
 
-        return $this->redirectToRoute('app_tag_index', [], Response::HTTP_SEE_OTHER);
+            // Supprimer la catégorie
+            if ($this->isCsrfTokenValid('delete' . $tag->getId(), $request->getPayload()->getString('_token'))) {
+                $entityManager->remove($tag);
+                $entityManager->flush();
+            }
+
+            return $this->redirectToRoute('app_tag_index', [], Response::HTTP_SEE_OTHER);
+        }
     }
 }
